ecal trig tower det id methods constexpr
